@@ -12,7 +12,9 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 import 'package:flutter_file_dialog/flutter_file_dialog.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(MaterialApp(
+      home: MyApp(),
+    ));
 
 class MyApp extends StatefulWidget {
   @override
@@ -24,11 +26,15 @@ class _MyAppState extends State<MyApp> {
   OpenFileDialogType _dialogType = OpenFileDialogType.image;
   SourceType _sourceType = SourceType.photoLibrary;
   bool _allowEditing = false;
-  File _currentFile;
-  String _savedFilePath;
+  File? _currentFile;
+  String? _savedFilePath;
   bool _localOnly = false;
   bool _copyFileToCacheDir = true;
-  String _pickedFilePath;
+  String? _pickedFilePath;
+
+  DirectoryLocation? _pickedDirecotry;
+  Future<bool> _isPickDirectorySupported =
+      FlutterFileDialog.isPickDirectorySupported();
 
   @override
   void initState() {
@@ -37,110 +43,133 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('FlutterFileDialog test app'),
-        ),
-        body: SingleChildScrollView(
-          padding: const EdgeInsets.symmetric(horizontal: 8),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              Padding(
-                padding: const EdgeInsets.only(top: 24),
-                child: CupertinoSegmentedControl<OpenFileDialogType>(
-                  children: const {
-                    OpenFileDialogType.document: Text("document"),
-                    OpenFileDialogType.image: Text("image"),
-                  },
-                  groupValue: _dialogType,
-                  onValueChanged: (value) =>
-                      setState(() => _dialogType = value),
-                ),
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('FlutterFileDialog test app'),
+      ),
+      body: SingleChildScrollView(
+        padding: const EdgeInsets.symmetric(horizontal: 8),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Padding(
+              padding: const EdgeInsets.only(top: 24),
+              child: CupertinoSegmentedControl<OpenFileDialogType>(
+                children: const {
+                  OpenFileDialogType.document: Text("document"),
+                  OpenFileDialogType.image: Text("image"),
+                },
+                groupValue: _dialogType,
+                onValueChanged: (value) => setState(() => _dialogType = value),
               ),
-              Padding(
-                padding: const EdgeInsets.only(top: 24),
-                child: CupertinoSegmentedControl<SourceType>(
-                  children: const {
-                    SourceType.photoLibrary: Text("photoLibrary"),
-                    SourceType.savedPhotosAlbum: Text("savedPhotosAlbum"),
-                    SourceType.camera: Text("camera"),
-                  },
-                  groupValue: _sourceType,
-                  onValueChanged: (value) =>
-                      setState(() => _sourceType = value),
-                ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 24),
+              child: CupertinoSegmentedControl<SourceType>(
+                children: const {
+                  SourceType.photoLibrary: Text("photoLibrary"),
+                  SourceType.savedPhotosAlbum: Text("savedPhotosAlbum"),
+                  SourceType.camera: Text("camera"),
+                },
+                groupValue: _sourceType,
+                onValueChanged: (value) => setState(() => _sourceType = value),
               ),
-              if (_dialogType == OpenFileDialogType.image)
-                Padding(
-                  padding: const EdgeInsets.only(top: 24),
-                  child: CheckboxListTile(
-                    title: const Text("Allow editing"),
-                    value: _allowEditing,
-                    onChanged: (v) => setState(() => _allowEditing = v),
-                  ),
-                ),
+            ),
+            if (_dialogType == OpenFileDialogType.image)
               Padding(
                 padding: const EdgeInsets.only(top: 24),
                 child: CheckboxListTile(
-                  title: const Text("Copy file to cache dir"),
-                  value: _copyFileToCacheDir,
-                  onChanged: (v) => setState(() => _copyFileToCacheDir = v),
+                  title: const Text("Allow editing"),
+                  value: _allowEditing,
+                  onChanged: (v) => setState(() => _allowEditing = v == true),
                 ),
               ),
+            Padding(
+              padding: const EdgeInsets.only(top: 24),
+              child: CheckboxListTile(
+                title: const Text("Copy file to cache dir"),
+                value: _copyFileToCacheDir,
+                onChanged: (v) =>
+                    setState(() => _copyFileToCacheDir = v == true),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 24),
+              child: ElevatedButton(
+                onPressed: _pickFile,
+                child: const Text('Pick file'),
+              ),
+            ),
+            FutureBuilder<bool>(
+              future: _isPickDirectorySupported,
+              builder: (context, snapshot) {
+                if (!snapshot.hasData || !snapshot.data!) return Container();
+                return Padding(
+                  padding: const EdgeInsets.only(top: 24),
+                  child: ElevatedButton(
+                    onPressed: _pickDirectory,
+                    child: const Text('Pick directory'),
+                  ),
+                );
+              },
+            ),
+            if (_pickedDirecotry != null)
+              Padding(
+                padding: const EdgeInsets.only(top: 24),
+                child: Text(_pickedDirecotry!.toString()),
+              ),
+            if (_pickedDirecotry != null && _currentFile != null)
               Padding(
                 padding: const EdgeInsets.only(top: 24),
                 child: ElevatedButton(
-                  onPressed: _pickFile,
-                  child: const Text('Pick file'),
+                  onPressed: _saveFileToDirectory,
+                  child: const Text('Save File to directory'),
                 ),
               ),
-              if (_pickedFilePath?.isNotEmpty == true)
-                Padding(
-                  padding: const EdgeInsets.only(top: 24),
-                  child: Text(_pickedFilePath),
+            if (_pickedFilePath?.isNotEmpty == true)
+              Padding(
+                padding: const EdgeInsets.only(top: 24),
+                child: Text(_pickedFilePath!),
+              ),
+            const SizedBox(
+              height: 24,
+            ),
+            if (_currentFile?.existsSync() == true) ...[
+              Text(
+                  "${(_currentFile!.lengthSync() / 1024.0).toStringAsFixed(1)} KB"),
+              SizedBox(
+                width: 200,
+                height: 200,
+                child: Image.file(
+                  _currentFile!,
                 ),
-              const SizedBox(
-                height: 24,
-              ),
-              if (_currentFile?.existsSync() == true) ...[
-                Text(
-                    "${(_currentFile.lengthSync() / 1024.0).toStringAsFixed(1)} KB"),
-                SizedBox(
-                  width: 200,
-                  height: 200,
-                  child: Image.file(
-                    _currentFile,
-                  ),
-                ),
-              ],
-              ElevatedButton(
-                onPressed: _currentFile == null ? null : () => _saveFile(false),
-                child: const Text('Save file'),
-              ),
-              ElevatedButton(
-                onPressed: _currentFile == null ? null : () => _saveFile(true),
-                child: const Text('Save file from data'),
-              ),
-              Text(_savedFilePath ?? "-"),
-              if (_isBusy) const CircularProgressIndicator(),
-              CheckboxListTile(
-                title: const Text("Local only"),
-                value: _localOnly,
-                onChanged: (v) => setState(() {
-                  _localOnly = v;
-                }),
               ),
             ],
-          ),
+            ElevatedButton(
+              onPressed: _currentFile == null ? null : () => _saveFile(false),
+              child: const Text('Save file'),
+            ),
+            ElevatedButton(
+              onPressed: _currentFile == null ? null : () => _saveFile(true),
+              child: const Text('Save file from data'),
+            ),
+            Text(_savedFilePath ?? "-"),
+            if (_isBusy) const CircularProgressIndicator(),
+            CheckboxListTile(
+              title: const Text("Local only"),
+              value: _localOnly,
+              onChanged: (v) => setState(() {
+                _localOnly = v == true;
+              }),
+            ),
+          ],
         ),
       ),
     );
   }
 
   Future<void> _pickFile() async {
-    String result;
+    String? result;
     try {
       setState(() {
         _isBusy = true;
@@ -170,15 +199,66 @@ class _MyAppState extends State<MyApp> {
     }
   }
 
+  Future<void> _pickDirectory() async {
+    _pickedDirecotry = (await FlutterFileDialog.pickDirectory());
+    setState(() {});
+  }
+
+  Future<void> _saveFileToDirectory() async {
+    String ext = _currentFile!.path.split('.').last;
+    if (ext == 'jpg') ext = 'jpeg';
+
+    final fileData = _currentFile!.readAsBytesSync();
+    final mimeType = 'image/$ext';
+    final newFileName = 'abc123.$ext';
+
+    FlutterFileDialog.saveFileToDirectory(
+      directory: _pickedDirecotry!,
+      data: fileData,
+      mimeType: mimeType,
+      fileName: newFileName,
+      replace: false,
+      onFileExists: () async {
+        await showDialog<void>(
+          context: context,
+          builder: (BuildContext context) {
+            return SimpleDialog(
+              title: const Text('File already exists'),
+              children: <Widget>[
+                SimpleDialogOption(
+                  child: const Text('Cancel'),
+                  onPressed: () => Navigator.pop(context),
+                ),
+                SimpleDialogOption(
+                  child: const Text('Replace'),
+                  onPressed: () {
+                    Navigator.pop(context);
+                    FlutterFileDialog.saveFileToDirectory(
+                      directory: _pickedDirecotry!,
+                      data: fileData,
+                      mimeType: mimeType,
+                      fileName: newFileName,
+                      replace: true,
+                    );
+                  },
+                ),
+              ],
+            );
+          },
+        );
+      },
+    );
+  }
+
   Future<void> _saveFile(bool useData) async {
-    String result;
+    String? result;
     try {
       setState(() {
         _isBusy = true;
       });
-      final data = useData ? await _currentFile.readAsBytes() : null;
+      final data = useData ? await _currentFile!.readAsBytes() : null;
       final params = SaveFileDialogParams(
-          sourceFilePath: useData ? null : _currentFile.path,
+          sourceFilePath: useData ? null : _currentFile!.path,
           data: data,
           localOnly: _localOnly,
           fileName: useData ? "untitled" : null);
